+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - **app:general:** re-attach renderer worker metrics to footer
 - **app:backend:** fix unstable directory listing of FSA API in chromium 122 & 123 through caching
 - **app:frontend:** fix an issue that caused color updates triggered through the author panel to not be applied properly
+- **app:backend:** ensure `maestro.updateQuery` only triggers necessary updates
 
 ---